Temporary fix for valid characters error
When adding a url, the invalid code characters message appeared even if the code was valid.

Seth inspected the form field and changed the pattern which killed the error (suggesting a problem with the regex).

Thus our quick fix was to make the Code Pattern accept all characters (.*) in the Add and Update Actions and Views.

We made these edits to the live container.